--- a/torchani/cuaev/aev.cu
+++ b/torchani/cuaev/aev.cu
@@ -133,7 +133,6 @@ __global__ void pairwiseDistanceSingleMolecule(
     torch::PackedTensorAccessor32<DataT, 3, torch::RestrictPtrTraits> pos_t,
     PairDist<DataT>* d_Rij,
     IndexT max_natoms_per_mol) {
-
   int mol_idx = 0;
   int natom_pairs = max_natoms_per_mol * max_natoms_per_mol;
   int i = blockIdx.x * blockDim.x + threadIdx.x;
@@ -900,26 +899,26 @@ Result cuaev_forward(
   const int block_size = 64;
 
   dim3 block(8, 8, 1);
-  if (n_molecules == 1){
+  if (n_molecules == 1) {
     int tileWidth = 32;
     int tilesPerRow = (max_natoms_per_mol + tileWidth - 1) / tileWidth;
     dim3 block(tileWidth, tileWidth, 1);
     dim3 grid(tilesPerRow, tilesPerRow, 1);
     pairwiseDistanceSingleMolecule<<<grid, block, 0, stream>>>(
-      species_t.packed_accessor32<int, 2, torch::RestrictPtrTraits>(),
-      coordinates_t.packed_accessor32<float, 3, torch::RestrictPtrTraits>(),
-      d_Rij,
-      max_natoms_per_mol);
-  }else{
+        species_t.packed_accessor32<int, 2, torch::RestrictPtrTraits>(),
+        coordinates_t.packed_accessor32<float, 3, torch::RestrictPtrTraits>(),
+        d_Rij,
+        max_natoms_per_mol);
+  } else {
     dim3 block(8, 8, 1);
     // Compute pairwise distance (Rij) for all atom pairs in a molecule
     // maximum 4096 atoms, which needs 49152 byte (48 kb) of shared memory
     // TODO: the kernel is not optimized for batched huge molecule (max_natoms_per_mol > 1000)
     pairwiseDistance<<<n_molecules, block, sizeof(float) * max_natoms_per_mol * 3, stream>>>(
-      species_t.packed_accessor32<int, 2, torch::RestrictPtrTraits>(),
-      coordinates_t.packed_accessor32<float, 3, torch::RestrictPtrTraits>(),
-      d_Rij,
-      max_natoms_per_mol);
+        species_t.packed_accessor32<int, 2, torch::RestrictPtrTraits>(),
+        coordinates_t.packed_accessor32<float, 3, torch::RestrictPtrTraits>(),
+        d_Rij,
+        max_natoms_per_mol);
   }
 
   // Extract Rijs that is needed for RadialAEV comptuation i.e. all the Rij <= Rcr