Related issue: #522